+++ fleetapp/ViewModels/FleetListViewModel.cs
@@ -85,8 +85,8 @@ namespace fleetapp.ViewModels
                     AssetType = data[0],
                     AssetModel = data[1],
                     FleetId = data[2],
-                    InitialAge = Int32.Parse(data[3]),
-                    FinalAge = Int32.Parse(data[4]),
+                    InitialAge = Decimal.Parse(data[3]),
+                    FinalAge = Decimal.Parse(data[4]),
                     Priority = 1
                 };
             });
